+++ Sources/Emitting/Options.swift
@@ -4,9 +4,14 @@ public extension PureYAML.Emitting {
         public static let `default` = Options()
 
         public var scalarStyle: ScalarStyle
+        public var collectionStyle: CollectionStyle
 
-        public init(scalarStyle: ScalarStyle = .quoted) {
+        public init(
+            scalarStyle: ScalarStyle = .quoted,
+            collectionStyle: CollectionStyle = .block,
+        ) {
             self.scalarStyle = scalarStyle
+            self.collectionStyle = collectionStyle
         }
     }
 }